Introduce SharedStorage
