+++ sdk/src/main/java/com/oursky/authgear/SharedStorage.kt
@@ -1,0 +1,11 @@
+package com.oursky.authgear
+
+interface SharedStorage {
+    fun setIDToken(namespace: String, idToken: String)
+    fun getIDToken(namespace: String): String?
+    fun deleteIDToken(namespace: String)
+
+    fun setDeviceSecret(namespace: String, deviceSecret: String)
+    fun getDeviceSecret(namespace: String): String?
+    fun deleteDeviceSecret(namespace: String)
+}
