+++ sdk/src/main/java/com/oursky/authgear/TokenStorage.kt
@@ -4,12 +4,4 @@ interface TokenStorage {
     fun setRefreshToken(namespace: String, refreshToken: String)
     fun getRefreshToken(namespace: String): String?
     fun deleteRefreshToken(namespace: String)
-
-    fun setIDToken(namespace: String, idToken: String)
-    fun getIDToken(namespace: String): String?
-    fun deleteIDToken(namespace: String)
-
-    fun setDeviceSecret(namespace: String, deviceSecret: String)
-    fun getDeviceSecret(namespace: String): String?
-    fun deleteDeviceSecret(namespace: String)
 }
